"Added JavaScript function for contact form submission and updated contact\_us.php to use it"
This diff includes changes to both contact\_us.php and script.js. In contact\_us.php, the submission button for the contact form has been updated to call the JavaScript function `ContactUsSend()` instead of directly submitting the form. Additionally, a new select element with the id "rid" has been added.

In script.js, the new function `ContactUsSend()` has been added. This function gets the values of the form fields in the contact form, creates a new FormData object with those values, and sends an HTTP POST request to 'contact-us-procces.php' with the FormData object as the request body. The function also handles the response from the server, updating the message and displaying a toast notification.