"Add product quantity field to cart and update add to cart functionality"
This commit adds a new column 'qty' to the 'cart' table in the database and updates the 'addToCart.php' file to include the product quantity in the insert query. Additionally, the 'product-details.php' file is updated to include an 'onclick' function for the add to cart button, which calls the new 'addToCart(i)' JavaScript function. This function gets the product quantity from an input field before adding the product to the cart. The 'script.js' file is also updated to include the new 'addToCart(i)' function, which displays an alert with the server response.