"Added session variable to store cart array; fixed form control syntax"
The code diff shows changes made to the 'my-cart.php' file. Specifically, it adds a new session variable called 'cart' and assigns it the value of the $array variable. The $array variable is populated with the results of a database query that fetches product information for items in a shopping cart.

Additionally, the diff shows a minor change in the syntax of a form control element, adding a space inside the input tag's closing angle bracket.